requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
streamlit
Pillow





